Added new data

+++ ESP32 code/src/main.cpp
@@ -1,26 +1,61 @@
 /*
-overflow of ESP.getCycleCount()
-asuming SD card is not full
-add rtc time of restart for compensation
-external RTC or accept ~5% drift
-*/
+Note:
+    overflow of ESP.getCycleCount()
+    assuming SD card is not full
+    assuming wifi is always connected
 
-/*
+To do:
+    rename file instead of deleting
+
 SD card file structure:
 /data
     - data.csv
-    - counter.txt  
+    - counter.txt 
+    - time_start.txt
+
+SD card -> ESP32
+ CS         5
+ SCK        18
+ MOSI       23
+ MISO       19
 */
+
+// option select
+#define wifi
+
 
 #include <Arduino.h>
 #include <SD.h>
-#include <sys/time.h>
+#include <time.h>
 
-#define input_pin 16
-#define timeout 24000 // for 100us at 240MHz clock
-#define baud_rate 115200
 
-//#define debug
+#ifdef wifi
+    #include <WiFi.h>
+    #include <NTPClient.h>
+    #include <WiFiUdp.h>
+    #include "Credentials.h"
+
+    #define UTC_OFFSET_IN_SECONDS 3600 // offset from Greenwitch time
+    #define NTP_TIME_UPDATE_INTERVAL 60000 // 1 minute
+
+    #ifdef CREDENTIALS_H
+        char ssid[] = WIFI_SSID;
+        char password[] = WIFI_PASSWORD;
+    #else
+        char ssid[] = "sample_ssid";
+        char password[] = "sample_password";
+    #endif
+
+    WiFiUDP ntpUDP;
+    NTPClient timeClient(ntpUDP, "cz.pool.ntp.org", UTC_OFFSET_IN_SECONDS, NTP_TIME_UPDATE_INTERVAL);
+#endif
+
+
+#define INPUT_PIN 16
+#define OUTPUT_PIN 33
+#define TIMEOUT 12000 // for 50 μs at 240MHz clock
+#define BAUD_RATE 115200
+
 
 volatile unsigned long time_a;
 volatile unsigned long time_b;
@@ -28,15 +63,16 @@ volatile unsigned long d_time;
 volatile bool interrupt_flag = false;
 int counter = 0;
 int serial_in_data = 0;
-struct timeval tv_now;
+
+
 
 void IRAM_ATTR isr()
 {
-    time_a = ESP.getCycleCount(); // Overflow every ~28s
+    time_a = ESP.getCycleCount(); // overflow every ~28s
     interrupt_flag = true;
 }
 
-//Source: https://github.com/markszabo/IRremoteESP8266/blob/master/src/IRutils.cpp#L48
+// Source: https://github.com/markszabo/IRremoteESP8266/blob/master/src/IRutils.cpp#L48
 String uint64_to_string(uint64_t input)
 {
     String result = "";
@@ -157,11 +193,12 @@ void delete_file(fs::FS &fs, const char *path)
 
 void setup()
 {
-    // input_pin setup
-    pinMode(input_pin, INPUT);
-    attachInterrupt(input_pin, isr, RISING);
+    // IO pins setup
+    pinMode(INPUT_PIN, INPUT);
+    attachInterrupt(INPUT_PIN, isr, RISING);
+    pinMode(OUTPUT_PIN, OUTPUT);
 
-    Serial.begin(baud_rate);
+    Serial.begin(BAUD_RATE);
 
     // SD card setup
     if (!SD.begin())
@@ -181,7 +218,7 @@ void setup()
         createDir(SD, "/data");
     }
 
-    //load counter value from SD card
+    // load counter value from SD card
     if (SD.exists("/data/counter.txt"))
     {
         File file = SD.open("/data/counter.txt");
@@ -197,35 +234,63 @@ void setup()
         Serial.println("Counter value loaded successfully!");
     }
 
-    Serial.println("Initialized!");
+    #ifdef wifi
+    WiFi.begin(ssid, password);
+
+    while(WiFi.status() != WL_CONNECTED)
+    {
+        delay(500);
+        Serial.print(".");
+    }
+    timeClient.begin();
+    timeClient.update();
+    delay(500);
+    Serial.println("Started at: " + String(timeClient.getFormattedTime()));
+    
+    // save startup time
+    write_file(SD, "/data/startup_time.txt", (String(timeClient.getEpochTime())).c_str());
+    #endif
 }
 
-#ifdef debug
+
+
 void loop()
 {
-    /*
-    d_time = 123456789;
-    Serial.println(d_time);
-
-    if (SD.exists("/data/data.csv"))
+    if (interrupt_flag)
     {
-        append_file(SD, "/data/data.csv", (String(d_time) + "," + uint64_to_string(esp_timer_get_time()) + ",").c_str());
-    }
-    else
-    {
-        write_file(SD, "/data/data.csv", (String(d_time) + "," + uint64_to_string(esp_timer_get_time()) + ",").c_str());
-    }
-    counter++;
+        time_b = time_a;
+        time_a = 0;
 
-    //save counter value to SD card
-    if (SD.exists("/data/counter.txt"))
-    {
-        delete_file(SD, "/data/counter.txt");
-    }
-    write_file(SD, "/data/counter.txt", (String(counter)).c_str());
+        while (ESP.getCycleCount() < time_b + TIMEOUT)
+        {
+            if (time_a != 0)
+            {
+                digitalWrite(OUTPUT_PIN, HIGH);
+                d_time = time_a - time_b;
+                Serial.println(d_time);
 
-    d_time++;
-    delay(5000);*/
+                if (SD.exists("/data/data.csv"))
+                {
+                    append_file(SD, "/data/data.csv", ("\n" + String(d_time) + "," + uint64_to_string(esp_timer_get_time()) + "," + String(timeClient.getEpochTime())).c_str());
+                }
+                else
+                {
+                    write_file(SD, "/data/data.csv", (String(d_time) + "," + uint64_to_string(esp_timer_get_time()) + "," + String(timeClient.getEpochTime())).c_str());
+                }
+                counter++;
+
+                // save counter value to SD card
+                if (SD.exists("/data/counter.txt"))
+                {
+                    delete_file(SD, "/data/counter.txt");
+                }
+                write_file(SD, "/data/counter.txt", (String(counter)).c_str());
+                digitalWrite(OUTPUT_PIN, LOW);
+                interrupt_flag = false; // <------------ too late?
+                break;
+            }
+        }
+    }
 
     if (Serial.available())
     {
@@ -233,21 +298,24 @@ void loop()
         serial_in_data -= 48;
         Serial.println(serial_in_data);
         if (serial_in_data == 1)
-        { //print data
+        { // print data
             Serial.println("Printing data:");
             readFile(SD, "/data/data.csv");
+            Serial.print("\n");
         }
         else if (serial_in_data == 2)
-        { //print counter
-            Serial.println(counter);
+        { // print counter
+            Serial.println("Count is: " + String(counter));
         }
         else if (serial_in_data == 3)
-        { //print time since boot
-            Serial.println("time_since_boot:" + uint64_to_string(esp_timer_get_time()) + "us");
+        { // print time since boot
+            Serial.println("Time_since_boot is : " + uint64_to_string(esp_timer_get_time()) + "us");
+            // print time
+            Serial.println("Current epoch time + offset is: " + String(timeClient.getEpochTime()) + " aka time: " + String(timeClient.getFormattedTime()));
         }
         else if (serial_in_data == 8)
-        { // reset counter value and save to SD card
-
+        { // reset counter value
+            // zero counter value and save to SD card
             if (SD.exists("/data/counter.txt"))
             {
                 delete_file(SD, "/data/counter.txt");
@@ -263,72 +331,11 @@ void loop()
             Serial.println("Restarting");
             ESP.restart();
         }
-        else
+        else if (serial_in_data == -6)
         {
-            Serial.println("Unknown command");
-        }
-        serial_in_data = 0;
-    }
-}
-#else
-void loop()
-{
-    if (interrupt_flag)
-    {
-        time_b = time_a;
-        time_a = 0;
+            Serial.println("Deleting data!");
+            delete_file(SD, "/data/data.csv");
 
-        while (ESP.getCycleCount() < time_b + timeout)
-        {
-            if (time_a != 0)
-            {
-                d_time = time_a - time_b;
-                Serial.println(d_time);
-
-                if (SD.exists("/data/data.csv"))
-                {
-                    append_file(SD, "/data/data.csv", (String(d_time) + "," + uint64_to_string(esp_timer_get_time()) + ",").c_str());
-                }
-                else
-                {
-                    write_file(SD, "/data/data.csv", (String(d_time) + "," + uint64_to_string(esp_timer_get_time()) + ",").c_str());
-                }
-                counter++;
-
-                //save counter value to SD card
-                if (SD.exists("/data/counter.txt"))
-                {
-                    delete_file(SD, "/data/counter.txt");
-                }
-                write_file(SD, "/data/counter.txt", (String(counter)).c_str());
-
-                interrupt_flag = false; // <------------ too late?
-                break;
-            }
-        }
-    }
-
-    if (Serial.available())
-    {
-        serial_in_data = Serial.read();
-        serial_in_data -= 48;
-        Serial.println(serial_in_data);
-        if (serial_in_data == 1)
-        { //print data <------------------------
-            Serial.println("Printing data:");
-            readFile(SD, "/data/data.csv");
-        }
-        else if (serial_in_data == 2)
-        { //print counter
-            Serial.println(counter);
-        }
-        else if (serial_in_data == 3)
-        { //print time since boot
-            Serial.println("time_since_boot:" + uint64_to_string(esp_timer_get_time()) + "us");
-        }
-        else if (serial_in_data == 8)
-        { //reset counter values
-            //zero counter value and save to SD card
             if (SD.exists("/data/counter.txt"))
             {
                 delete_file(SD, "/data/counter.txt");
@@ -339,11 +346,6 @@ void loop()
 
             Serial.println("counter value reset!");
         }
-        else if (serial_in_data == 9)
-        { //restart
-            Serial.println("Restarting");
-            ESP.restart();
-        }
         else
         {
             Serial.println("Unknown command");
@@ -351,4 +353,3 @@ void loop()
         serial_in_data = 0;
     }
 }
-#endif
